Refactor styling for MediumSizedCompanies and StartUps components
- Updated hero section background and text colors for MediumSizedCompanies.
- Changed button styles to improve visibility and consistency in MediumSizedCompanies.
- Modified contact section background and text styles in MediumSizedCompanies.
- Adjusted section styles for StartUps component to match new design standards.
- Enhanced button styles in StartUps for better user interaction.
